+++ Backend/WebApis/Dtos/SkillsDto.cs
@@ -3,6 +3,6 @@
     public class SkillsDto
     {
         public string Name { get; set; } = string.Empty;
-        public int Experience { get; set; } // in years
+        public Decimal Experience { get; set; } // in years
     }
 }
